+++ src/fs/BrowserFS.d.ts
