+++ src/language/openscad-register-language.ts
@@ -1,7 +1,7 @@
 // Portions of this file are Copyright 2021 Google LLC, and licensed under GPL2+. See COPYING.
 
 import loader from '@monaco-editor/loader';
-import { ZipArchives } from '../zip-archives';
+import { ZipArchives } from '../fs/zip-archives';
 import { buildOpenSCADCompletionItemProvider } from './openscad-completions';
 import openscadLanguage from './openscad-language';
 // import jsLanguage from 'monaco-editor/esm/vs/basic-languages/javascript/javascript.contribution'
